Separated global api to two apis.
