+++ frontend/src/pages/HomePage.js
@@ -6,7 +6,7 @@ import ObjectiveCard from '../components/ObjectiveCard';
 import ActionForm from '../components/ActionForm';
 import ActionCard from '../components/ActionCard';
 
-import { getActionsForUser, getObjectivesForUser, deleteObjective, finishObjective, resetObjective, deleteAction, finishAction, resetAction, addCommentToAction } from '../services/api';
+import { actionsApi, objectivesApi } from '../services/api/index';
 
 const HomePage = () => {
   const [actions, setActions] = useState([]);
@@ -19,7 +19,7 @@ const HomePage = () => {
 
   const fetchActions = async () => {
     try {
-      const fetchedActions = await getActionsForUser('root') // To change
+      const fetchedActions = await actionsApi.getActionsForUser('root') // To change
       setActions(fetchedActions);
     } catch (error) {
       console.error('Error fetching actions:', error);
@@ -28,7 +28,7 @@ const HomePage = () => {
 
   const fetchObjectives = async () => {
     try {
-      const fetchedObjectives = await getObjectivesForUser('root') // To change
+      const fetchedObjectives = await objectivesApi.getObjectivesForUser('root') // To change
       setObjectives(fetchedObjectives);
     } catch (error) {
       console.error('Error fetching objectives:', error);
@@ -37,7 +37,7 @@ const HomePage = () => {
 
   const handleDeleteObjective = async (objectiveId) => {
     try {
-      await deleteObjective(objectiveId);
+      await objectivesApi.deleteObjective(objectiveId);
       setObjectives(objectives.filter((objective) => objective.id !== objectiveId));
     } catch (error) {
       console.error('Error deleting objective:', error);
@@ -46,7 +46,7 @@ const HomePage = () => {
 
   const handleResetObjective = async (objectiveId) => {
     try {
-      await resetObjective(objectiveId);
+      await objectivesApi.resetObjective(objectiveId);
       fetchObjectives();
     } catch (error) {
       console.error('Error resetting objective:', error);
@@ -55,7 +55,7 @@ const HomePage = () => {
 
   const handleFinishObjective = async (objectiveId, realDuration) => {
     try {
-      await finishObjective(objectiveId, realDuration);
+      await objectivesApi.finishObjective(objectiveId, realDuration);
       fetchObjectives();
     } catch (error) {
       console.error('Error finishing objective:', error);
@@ -64,7 +64,7 @@ const HomePage = () => {
 
   const handleDeleteAction = async (actionId) => {
     try {
-      await deleteAction(actionId);
+      await actionsApi.deleteAction(actionId);
       setActions(actions.filter((action) => action.id !== actionId));
     } catch (error) {
       console.error('Error deleting action:', error);
@@ -73,7 +73,7 @@ const HomePage = () => {
 
   const handleResetAction = async (actionId) => {
     try {
-      await resetAction(actionId);
+      await actionsApi.resetAction(actionId);
       fetchActions();
     } catch (error) {
       console.error('Error resetting action:', error);
@@ -82,7 +82,7 @@ const HomePage = () => {
 
   const handleFinishAction = async (actionId, realDuration) => {
     try {
-      await finishAction(actionId, realDuration);
+      await actionsApi.finishAction(actionId, realDuration);
       fetchActions();
     } catch (error) {
       console.error('Error finishing action:', error);
@@ -91,7 +91,7 @@ const HomePage = () => {
 
   const handleSaveActionComment = async (id, comment) => {
     try {
-      await addCommentToAction(id, comment);
+      await actionsApi.addCommentToAction(id, comment);
       fetchActions();
     } catch (error) {
       console.error('Error saving comment:', error);
