+++ frontend/src/services/api/index.js
@@ -1,0 +1,7 @@
+import * as actionsApi from './actions';
+import * as objectivesApi from './objectives';
+
+export {
+    actionsApi,
+    objectivesApi,
+};
